added some conditions for updating wb_menu_category

+++ apps/wb_checker/wb_menu_categories.py
@@ -189,6 +189,6 @@ class TopWBProductMenuCategoryUpdater():
     def run(self):
         author_object = CustomUser.objects.get(username='coldtema')
         for category in self.all_categories:
-            if category.shard_key != 'blackhole':
+            if category.shard_key != 'blackhole' and category.query != None and category.shard_key != None and category.shard_key != '':
                 TopWBProduct.objects.filter(source='CATEGORY', menu_category=category).delete()
                 MenuCategory(f'https://www.wildberries.ru{category.main_url}', author_object).run()
